webapp-cg/requirements.txt
Flask>=2.0
Werkzeug>=2.2
pycryptodome
Pillow
qrcode
